// Angular-FE/src/app/Interface/car.interface.ts
export interface Car {
  id: number;
  make: string;
  model: string;
  year: number;
  color: string;
  mileage: number;
  price: number;
  fuelType: string;
  transmission: string;
  engine: string;
  horsepower: number;
  features: string[];   // Features as an array of strings
  owners: number;       // Number of previous owners
  image: string;
}